refactor: account API
