+++ index.js
@@ -3,7 +3,7 @@ const pulumi = require("@pulumi/pulumi");
 const aws = require("@pulumi/aws");
 const awsx = require("@pulumi/awsx");
 
-const { mainnet, polygon, bsc, getAllProjects, getProject, getAllEvents, getEvent, generateProof } = require("./routes")
+const { mainnet, polygon, bsc, getAllProjects, getProject, getAllEvents, getEvent, generateProof, getAccount, createAccount } = require("./routes")
 
 const Headers = {
     "Access-Control-Allow-Headers": "Content-Type",
@@ -86,109 +86,13 @@ const angpowApi = new awsx.apigateway.API("angpow-api", {
             eventHandler: async (event) => await bsc(event)
         },
         {
-            method: "GET", path: "/account/{proxy+}", eventHandler: async (event) => {
-
-                if (event && event.pathParameters) {
-                    const client = new aws.sdk.DynamoDB.DocumentClient()
-                    const accountId = event.pathParameters.proxy
-
-                    const params = {
-                        TableName: dataTable.name.get(),
-                        Key: {
-                            "key": "account",
-                            "value": accountId
-                        }
-                    };
-
-                    const { Item } = await client.get(params).promise()
-
-                    if (Item) {
-                        return {
-                            statusCode: 200,
-                            headers: Headers,
-                            body: JSON.stringify({
-                                status: "ok",
-                                account: accountId,
-                                disabled: Item.disabled
-                            }),
-                        }
-                    }
-                }
-
-                return {
-                    statusCode: 200,
-                    headers: Headers,
-                    body: JSON.stringify({
-                        status: "error",
-                        message: "Invalid ID"
-                    }),
-                }
-            }
+            method: "GET",
+            path: "/account/{proxy+}",
+            eventHandler: async (event) => await getAccount(event, dataTable.name.get())
         },
         // FIXME: sign the message and verify before add a new record / use POST method and fix CORS issues
         {
-            method: "GET", path: "/accountUpdate/{proxy+}", eventHandler: async (event) => {
-
-                console.log("Incoming event --> ", event)
-
-                if (event && event.pathParameters) {
-
-                    // const client = new aws.sdk.DynamoDB.DocumentClient()
-                    const base64String = event.pathParameters.proxy
-
-                    const buff = Buffer.from(base64String, "base64");
-                    const eventBodyStr = buff.toString('UTF-8');
-                    const eventBody = JSON.parse(eventBodyStr);
-
-                    console.log("Receiving payload : ", eventBody)
-
-                    // example payload
-                    // {
-                    //     username : pisuthd.nft,
-                    //     address : xxxx,
-                    //     disabled : false,
-                    //     email : pisuth@tamago.finance
-                    // }
-
-                    if (eventBody && eventBody.username) {
-
-                        const params = {
-                            TableName: dataTable.name.get(),
-                            Item: {
-                                ...eventBody,
-                                "key": "account",
-                                "value": eventBody.username
-                            }
-                        }
-
-                        console.log("saving : ", params)
-
-                        const client = new aws.sdk.DynamoDB.DocumentClient()
-
-                        await client.put(params).promise()
-
-                        return {
-                            statusCode: 200,
-                            headers: Headers,
-                            body: JSON.stringify({
-                                "status": "ok",
-                                "username": eventBody.username
-                            }),
-                        }
-
-                    }
-
-                }
-
-                return {
-                    statusCode: 200,
-                    headers: Headers,
-                    body: JSON.stringify({
-                        status: "error",
-                        message: "Invalid Input"
-                    }),
-                }
-            },
+            method: "GET", path: "/accountUpdate/{proxy+}", eventHandler: async (event) => await createAccount(event, dataTable.name.get()),
         }
     ]
 });
@@ -220,6 +124,14 @@ const LuckboxApi = new awsx.apigateway.API("luckbox-api", {
             path: "/events/proof/{proxy+}",
             eventHandler: async (event) => await generateProof(event, { dataTable: dataTable.name.get(), projectTable: projectTable.name.get() })
         },
+        {
+            method: "GET",
+            path: "/accounts/{proxy+}",
+            eventHandler: async (event) => await getAccount(event, dataTable.name.get())
+        },
+        {
+            method: "GET", path: "/createAccount/{proxy+}", eventHandler: async (event) => await createAccount(event, dataTable.name.get()),
+        }
         // {
         //     method: "GET",
         //     path: "/orders",
