+++ routes/index.js
@@ -1,7 +1,7 @@
 const { mainnet , polygon, bsc } = require("./angpow")
 const { getAllProjects, getProject } = require("./projects")
 const { getAllEvents, getEvent, generateProof } = require("./events")
-
+const { getAccount, createAccount } = require("./account")
 
 const headers = {
     "Access-Control-Allow-Headers": "Content-Type",
@@ -18,5 +18,7 @@ module.exports = {
     getProject,
     getAllEvents,
     getEvent,
-    generateProof
+    generateProof,
+    getAccount,
+    createAccount
 }
